Zero velocity fix (#109)
* zero velocity fix
* fix trajectory_angle
* test_trajectory output extended
* cleaned up cython code
