+++ py_ballisticcalc.exts/py_ballisticcalc_exts/early_bind_config.pxd
@@ -1,4 +1,4 @@
-cdef struct _Config:
+cdef struct _ConfigStruct:
     bint use_powder_sensitivity
     double max_calc_step_size_feet
     double cZeroFindingAccuracy
@@ -8,4 +8,4 @@ cdef struct _Config:
     double cGravityConstant
     double cMinimumAltitude
 
-cdef _Config _early_bind_config(object config)
+cdef _ConfigStruct _early_bind_config(object config)
