+++ py_ballisticcalc.exts/py_ballisticcalc_exts/early_bind_config.pyx
@@ -1,10 +1,10 @@
 from cython cimport final
-from py_ballisticcalc_exts.early_bind_config cimport _Config
+# from py_ballisticcalc_exts.early_bind_config cimport _ConfigStruct
 
 
 @final
-cdef _Config _early_bind_config(object config):
-    return _Config(
+cdef _ConfigStruct _early_bind_config(object config):
+    return _ConfigStruct(
         config.use_powder_sensitivity,
         config.max_calc_step_size_feet,
         config.cZeroFindingAccuracy,
