+++ tests/test_computer.py
@@ -201,7 +201,7 @@ class TestComputer(unittest.TestCase):
         self.assertLess(t.trajectory[0].velocity, self.baseline_trajectory[0].velocity)
         set_global_use_powder_sensitivity(previous)
 
-    @unittest.skip("Raises ZeroDivisionError")
+    # @unittest.skip("Raises ZeroDivisionError")
     def test_zero_velocity(self):
         tdm = DragModel(self.dm.BC + 0.5, self.dm.drag_table, self.dm.weight, self.dm.diameter, self.dm.length)
         slick = Ammo(tdm, 0)
